[asimage] Port to Win64 (replace Long_t by Longptr_t + pointer formatting)

+++ graf2d/asimage/src/libAfterImage/asim_afterbase.h
@@ -303,7 +303,7 @@ typedef union ASHashableValue
 }
 ASHashableValue;
 #else
-typedef unsigned long ASHashableValue;
+typedef uintptr_t ASHashableValue;
 #endif
 
 typedef union ASHashData
@@ -323,7 +323,7 @@ typedef union ASHashData
 	CARD8  c8 ;
 }ASHashData;
 
-#define AS_HASHABLE(v)  ((ASHashableValue)((unsigned long)(v)))
+#define AS_HASHABLE(v)  ((ASHashableValue)((uintptr_t)(v)))
 
 typedef struct ASHashItem
 {
